Sync branches for CI usage

+++ WinImageBuilder.psm1
@@ -437,6 +437,15 @@ function Is-IsoFile {
     return ([System.IO.Path]::GetExtension($FilePath) -eq ".iso")
 }
 
+function Is-ServerInstalationType {
+    Param(
+        [parameter(Mandatory=$true)]
+        [object]$image
+    )
+    return ($image.ImageInstallationType -in @("Server", "Server Core"))
+}
+
+
 function Add-VirtIODrivers {
     Param(
         [parameter(Mandatory=$true)]
@@ -468,20 +477,20 @@ function Add-VirtIODrivers {
     if ($image.ImageVersion.Major -eq 6 -and $image.ImageVersion.Minor -eq 0) {
         $virtioVer = "2k8"
     } elseif ($image.ImageVersion.Major -eq 6 -and $image.ImageVersion.Minor -eq 1) {
-        if ($image.ImageInstallationType -eq "Server") {
+        if (Is-ServerInstalationType $image) {
             $virtioVer = "2k8r2"
         } else {
             $virtioVer = "w7"
         }
     } elseif ($image.ImageVersion.Major -eq 6 -and $image.ImageVersion.Minor -eq 2) {
-        if ($image.ImageInstallationType -eq "Server") {
+        if (Is-ServerInstalationType $image) {
             $virtioVer = "2k12"
         } else {
             $virtioVer = "w8"
         }
     } elseif (($image.ImageVersion.Major -eq 6 -and $image.ImageVersion.Minor -ge 3) `
         -or $image.ImageVersion.Major -gt 6) {
-        if ($image.ImageInstallationType -eq "Server") {
+        if (Is-ServerInstalationType $image) {
             $virtioVer = "2k12R2"
         } else {
             $virtioVer = "w8.1"
@@ -635,8 +644,13 @@ function Resize-VHDImage {
         Write-Host "New partition size: $newSizeGB GB"
 
         if ($NewSize -gt $MinSize) {
+            $global:i = 0
+            $step = 100MB
             Execute-Retry {
-                Resize-Partition -DriveLetter $Drive -Size ($NewSize) -ErrorAction "Stop"
+                $sizeIncreased = ($NewSize + ($step * $global:i))
+                Write-Host "Size increased: $sizeIncreased"
+                $global:i = $global:i + 1
+                Resize-Partition -DriveLetter $Drive -Size $sizeIncreased -ErrorAction "Stop"
             }
         }
     }
@@ -687,6 +701,7 @@ function Check-Prerequisites {
 
 function GetOrCreate-Switch {
     $vmSwitches = Get-VMSwitch -SwitchType external
+    $vmswitch = $null
     if ($vmSwitches) {
         foreach ($i in $vmSwitches) {
             $name = $i.Name
@@ -786,10 +801,71 @@ function New-MaaSImage {
         [parameter(Mandatory=$false)]
         [string]$SwitchName,
         [parameter(Mandatory=$false)]
-        [switch]$Force=$false
+        [switch]$Force=$false,
+        [parameter(Mandatory=$false)]
+        [switch]$PurgeUpdates,
+        [parameter(Mandatory=$false)]
+        [switch]$DisableSwap
+    )
+        PROCESS
+    {
+        New-WindowsOnlineImage -Type "MAAS" -WimFilePath $WimFilePath -ImageName $ImageName `
+            -WindowsImagePath $MaaSImagePath -SizeBytes $SizeBytes -DiskLayout $DiskLayout `
+            -ProductKey $ProductKey -VirtIOISOPath $VirtIOISOPath -InstallUpdates:$InstallUpdates `
+            -AdministratorPassword $AdministratorPassword -PersistDriverInstall:$PersistDriverInstall `
+            -ExtraDriversPath $ExtraDriversPath -Memory $Memory -CpuCores $CpuCores `
+            -RunSysprep:$RunSysprep -SwitchName $SwitchName -Force:$Force -PurgeUpdates:$PurgeUpdates `
+            -DisableSwap:$DisableSwap
+    }
+}
+
+function New-WindowsOnlineImage {
+    [CmdletBinding()]
+    param
+    (
+        [parameter(Mandatory=$true, ValueFromPipeline=$true)]
+        [string]$WimFilePath = "D:\Sources\install.wim",
+        [parameter(Mandatory=$true)]
+        [string]$ImageName,
+        [parameter(Mandatory=$true)]
+        [string]$WindowsImagePath,
+        [parameter(Mandatory=$true)]
+        [Uint64]$SizeBytes,
+        [ValidateSet("BIOS", "UEFI", ignorecase=$false)]
+        [string]$DiskLayout = "BIOS",
+        [parameter(Mandatory=$false)]
+        [string]$ProductKey,
+        [parameter(Mandatory=$false)]
+        [string]$VirtIOISOPath,
+        [parameter(Mandatory=$false)]
+        [switch]$InstallUpdates,
+        [parameter(Mandatory=$false)]
+        [string]$AdministratorPassword = "Pa`$`$w0rd",
+        [array]$ExtraFeatures = @(),
+        [parameter(Mandatory=$false)]
+        [string]$ExtraDriversPath,
+        [parameter(Mandatory=$false)]
+        [switch]$PersistDriverInstall = $true,
+        [parameter(Mandatory=$false)]
+        [Uint64]$Memory=2GB,
+        [parameter(Mandatory=$false)]
+        [int]$CpuCores=1,
+        [parameter(Mandatory=$false)]
+        [switch]$RunSysprep=$true,
+        [parameter(Mandatory=$false)]
+        [string]$SwitchName,
+        [parameter(Mandatory=$false)]
+        [switch]$Force=$false,
+        [ValidateSet("MAAS", "KVM", "HYPER-V", ignorecase=$false)]
+        [string]$Type = "MAAS",
+        [parameter(Mandatory=$false)]
+        [switch]$PurgeUpdates,
+        [parameter(Mandatory=$false)]
+        [switch]$DisableSwap
     )
     PROCESS
     {
+        Write-Host ("Windows online image generation started at: {0}" -f @(Get-Date))
         Is-Administrator
         if (!$RunSysprep -and !$Force) {
             throw "You chose not to run sysprep.
@@ -825,15 +901,21 @@ function New-MaaSImage {
         }
 
         try {
-            $barePath = Get-PathWithoutExtension $MaaSImagePath
+            $barePath = Get-PathWithoutExtension $WindowsImagePath
             $VirtualDiskPath = $barePath + ".vhdx"
-            $RawImagePath = $barePath + ".img"
+            $InstallMaaSHooks = $false
+            if ($Type -eq "MAAS") {
+                $InstallMaaSHooks = $true
+            }
+            if ($Type -eq "KVM") {
+                $PersistDriverInstall = $false
+            }
             New-WindowsCloudImage -WimFilePath $WimFilePath -ImageName $ImageName `
                 -VirtualDiskPath $VirtualDiskPath -SizeBytes $SizeBytes -ProductKey $ProductKey `
                 -VirtIOISOPath $VirtIOISOPath -InstallUpdates:$InstallUpdates `
                 -AdministratorPassword $AdministratorPassword -PersistDriverInstall:$PersistDriverInstall `
-                -InstallMaaSHooks -ExtraFeatures $ExtraFeatures -ExtraDriversPath $ExtraDriversPath `
-                -DiskLayout $DiskLayout
+                -InstallMaaSHooks:$InstallMaaSHooks -ExtraFeatures $ExtraFeatures -ExtraDriversPath $ExtraDriversPath `
+                -DiskLayout $DiskLayout -PurgeUpdates:$PurgeUpdates -DisableSwap:$DisableSwap
 
             if ($RunSysprep) {
                 if($DiskLayout -eq "UEFI") {
@@ -842,22 +924,35 @@ function New-MaaSImage {
                     $generation = "1"
                 }
 
-                $Name = "MaaS-Sysprep" + (Get-Random)
+                $Name = "WindowsOnlineImage-Sysprep" + (Get-Random)
                 Run-Sysprep -Name $Name -Memory $Memory -VHDPath $VirtualDiskPath `
-                    -VMSwitch $switch.Name -CpuCores $CpuCores -Generation $generation
+                    -VMSwitch $switch.Name -CpuCores $CpuCores `
+                    -Generation $generation
             }
 
             Resize-VHDImage $VirtualDiskPath
 
-            Write-Host "Converting VHD to RAW"
-            Convert-VirtualDisk $VirtualDiskPath $RawImagePath "RAW"
-            Remove-Item -Force $VirtualDiskPath
-            Compress-Image $RawImagePath $MaaSImagePath
+            if ($Type -eq "MAAS") {
+                $RawImagePath = $barePath + ".img"
+                Write-Host "Converting VHD to RAW"
+                Convert-VirtualDisk $VirtualDiskPath $RawImagePath "RAW"
+                Remove-Item -Force $VirtualDiskPath
+                Compress-Image $RawImagePath $WindowsImagePath
+            }
+
+            if ($Type -eq "KVM") {
+                $Qcow2ImagePath = $barePath + ".qcow2"
+                Write-Host "Converting VHD to QCow2"
+                Convert-VirtualDisk $VirtualDiskPath $Qcow2ImagePath "qcow2"
+                Remove-Item -Force $VirtualDiskPath
+            }
         } catch {
-            Remove-Item -Force $MaaSImagePath* -ErrorAction SilentlyContinue
+            Remove-Item -Force $WindowsImagePath* -ErrorAction SilentlyContinue
             Throw
         }
+        Write-Host ("Windows online image generation finished at: {0}" -f @((Get-Date)))
     }
+
 }
 
 function New-WindowsCloudImage {
@@ -896,11 +991,16 @@ function New-WindowsCloudImage {
         [parameter(Mandatory=$false)]
         [switch]$InstallMaaSHooks,
         [parameter(Mandatory=$false)]
-        [string]$VirtIOBasePath
+        [string]$VirtIOBasePath,
+        [parameter(Mandatory=$false)]
+        [switch]$PurgeUpdates,
+        [parameter(Mandatory=$false)]
+        [switch]$DisableSwap
     )
 
     PROCESS
     {
+        Write-Host ("Image generation started at: {0}" -f @(Get-Date))
         Set-DotNetCWD
         Is-Administrator
 
@@ -930,6 +1030,8 @@ function New-WindowsCloudImage {
             $configValues = @{
                 "InstallUpdates"=$InstallUpdates;
                 "PersistDriverInstall"=$PersistDriverInstall;
+                "PurgeUpdates"=$PurgeUpdates;
+                "DisableSwap"=$DisableSwap;
             }
 
             Generate-UnattendXml $UnattendXmlPath $unattedXmlPath $image $ProductKey $AdministratorPassword
@@ -962,7 +1064,10 @@ function New-WindowsCloudImage {
             Convert-VirtualDisk $VHDPath $VirtualDiskPath $VirtualDiskFormat
             Remove-Item -Force $VHDPath
         }
+        Write-Host ("Image generation finished at: {0}" -f @(Get-Date))
     }
 }
 
-Export-ModuleMember New-WindowsCloudImage, Get-WimFileImagesInfo, New-MaaSImage, Resize-VHDImage
+Export-ModuleMember New-WindowsCloudImage, Get-WimFileImagesInfo, New-MaaSImage, Resize-VHDImage,
+    New-WindowsOnlineImage
+
